add NO

+++ DGL_model.py
@@ -610,9 +610,9 @@ class EGKN(torch.nn.Module):
 
     def forward(self, g, in_feat):
         h = in_feat
-        edge_index = torch.cat([g.edges()[0][None, :], g.edges()[1][None, :]], axis = 0)
+        edge_index = torch.cat([g.edges()[0][None, :], g.edges()[1][None, :]], axis = 0).type(torch.int64)
         coords_curr = g.ndata['feat'][:, :2].detach().clone()
-        edge_attr = torch.cat([g.edata['weight'][0][:, None], g.edata['slope'][0][:, None]], axis = 1)
+        edge_attr = torch.cat([g.edata['weight'][0][:, None], g.edata['slope'][0][:, None]], axis = 1).type(torch.float32)
 
         h = self.fc1(h)
         for k in range(self.depth):
